requirements-dev.txt
pytest
pytest-cov
pytest-homeassistant-custom-component
pytest-asyncio
bump-my-version==1.1.2
# From manifest
premium-bond-checker==1.0.2
